Updated Available Projects preamble text to include why join MME and scholarships
It would be good if we made an extra tab as this one is getting too long and has multiple subjects now. Could you make a new tab called "Why join?" and then put everything from the top of the Available projects page down to the end of the scholarships section. This will leave the Available projects page just to focus on the Research Opportunities